# feat: Aes algorithm hardware acceleration