@driftluo 2019-10-14
Because of  https://github.com/RustCrypto/stream-ciphers/issues/12, I think we should follow industry standards.

I will switch all encryption related libraries to openssl and ring to ensure compliance with industry standards.